fixes to G2P loading no more special cutoff for NR in TDL calc
